Move spatial trees into SpatialCollections
* This way they don't pollute the Collections namespace
